Add renewal of free membership
This change allows us to be able to remove members from our membership
list. This is necessary while we have a $0 membership fee, since members
would normally only be removed if they did not pay.

The intention is that we would email all members to renew their
membership by visiting the website 2 months before an election. Anyone
who did not respond would be removed from the membership.

They would still be able to reapply for membership at any time before
the election takes place and therefore vote in it.