Change the financial year to start on 1 October
This changes the start of the financial year from the 1st January to the
1st October while still keeping it 12 months.

This change is necessary since we are now required to have an auditor
create an oficial audit of our accounts for presentation at the AGM and
submission to Fair Trading.

Since our AGM is normally held at the end of January this does not leave
enough time to have a completed audit by the AGM especially in a time
when our accounts are in a large state of flux due to our largest
conference being on at the same time.