Reorganize timing of several events around financials
The goal here is to bring the checklists for both council and event
team into line with the more detailed information below, and also in
line with how the policy has actually been applied.

Key changes:

* In practice, event teams have been required to provide a budget
  before their subcommittee is approved. This is now reflected on the
  checklist.

* More clarity over what information needs to be provided to the
  council regarding Westpac accounts - the previous checklist only
  asked for names, which caused delays as people had to collect a
  second round of information.

* Move provision of budget templates to before the point where the
  team needs to prepare a budget.

* Move the discussions around expected profit to before the
  subcomittee is created, in line with current practice

* Perform due diligence on the proposed subcommittee members
  prior to forming the subcommittee

I think that most of this is supported by the explanatory text. For
instance, the draft proposal includes an initial budget, which is in
line with recent practice, and this is now reflected in the checklist.