Simplify conference budgeting by clarifying profit reporting
Currently, we ask conferences to return $11/head plus a negotiated
"surplus component".

However, our financial reports only report a single "profit" from the
conference; it's not clear from our reports how much of this covers LA
expenses and how much is genuine profit.

This change simplifies accounting and reporting by removing the
surplus component.

It is invisaged this would be complemented by:
 - A change from $11/head to 6% of total revenue as the fee LA charges
 - A change in reporting to show this as a fee returned to LA,
 reducing the reported "profit" by this amount

The "Budgeting" section already suggests that conference budgets
should "consider both direct concerns of the conference, and long-term
investments in the community". This change would make it clearer how
much profit is available for these long-term concerns.

It is anticipated that most events will want to prioritise these items
as "If we have space in the budget" items.