Initial import of the Subcommittee policy v3.
Import the Subcommittee policy v3 from the working document in gdocs to the
formal Linux Australia policy repository.  This completes the move to the
new version of the Subcommittee policy, which has been in operation since
early 2020.

The text is largely a verbatim copy of the working document.  However, some
minor grammatical errors noticed during the import process have been fixed.

It is possible that additional commits will be needed to fix any Markdown
rendering issues on the git host.