Improve inclusiveness in lists of characteristics in the Code of Conduct.
Following feedback and concern raised about the content of some documents
drawn on by the Linux Australia's Code Of Conduct[1], Council has decided to
include some clarifications to the code.

In two bullet points in the "What does that mean for me?" section, the Code
Of Conduct presents a list of examples without making it clear that it is a
non-exclusive list.  The words "not limited to" have been inserted to make
this explicit.

Council have also been made aware[2] that the omission of the "sex"
characteristic is problematic because it is not adequately covered by any of
the other characteristics in the lists given under "What does that mean for
me?".  Accordingly "sex" has been added in both cases.

The Linux Australia Council believe that these changes address the
concerns expressed in [2] as far as they apply to the Linux Australia
Code Of Conduct.  Council remains open to further feedback on issues
such as these and would welcome suggestions which address any remaining
issues.

[1] https://github.com/linuxaustralia/constitution_and_policies/blob/master/code_of_conduct.md
[2] http://files.au.adversary.org/files/django-advice-001.pdf