Import "Guidelines for Grant Submissions" from current PDF.
This document was prepared by the Linux Australia Council in 2021.  A PDF
was produced and linked to on the website but the text was never imported
into this repository as a primary reference.  This commit corrects the
oversight.

With this commit, the primary reference of the "Guidelines for Grant
Submissions" document is considered to be the copy in this repository.