Grant guidelines: try to clarify expectations for hardware purchases.
Over recent years there has been an increased interest in using Linux
Australia grants to fund hardware purchases in support of software
development activities.  When interfacing with hardware, software
development can often take unexpected turns which makes it difficult to
determine whether the exercise is considered a success at its conclusion.
To address this challenge and to ensure equitable access to the grants
across the Linux Australia membership, it is proposed that the information
in this commit be added to the grant guidelines.

This information will make it easier for the Linux Australia Council and the
wider membership to evaluate the effectiveness of these grants.  Where
applicable, it can also help determine whether additional future support for
the project, group or activity represents good value for Linux Australia,
its members and the Open Source community in general.