guidelines_for_grant_submissions.md: add report about funded hardware's future.
@lloy0076 suggested that when a grant includes funding for hardware, the
final report to Linux Australia ought to provide some visibility into how
that hardware might be used at the conclusion of the funded activities.
While it's difficult to be specific with the reporting requirements due to
the wide variety of hardware available, a general statement requesting
information about the hardware's future is possible.  This will increase
transparency around the acquired hardware.