Adopt the 2024-05-31 version of NSW Model constitution
When doing the Not-For-Profit self assessment [0] required by the ATO,
we encountered the following question:

    Question 3: Does the organisation have and follow clauses in its
    governing documents that prohibit the distribution of income or
    assets to members while it is operating and winding up?

As our current constitution lacks such a clause we answered "No", which
causes this to appear:

    The organisation can still self-assess as income tax exempt if it
    doesn't have these types of clauses in its governing documents,
    provided it has not distributed any assets or income to members.
    However, it has until 30 June 2025 to update its governing documents.
    Failure to do so will mean that it cannot self-assess as income tax
    exempt from 1 July 2024.

Our current constitution is based on the NSW Model constitution available
at the time.  The NSW Model constitution has evolved since then and now
does have the required clause.  This series of commits that adopts the new
NSW model constitution and customises it for Linux Australia, so that we
meet both the ATO and Australian Charities and Not-For-Profit commissions
requirement for Not-For-Profits.

This first commit imports the 2024-05-31 version of the NSW Model
constitution for organisations (the most recent at the time, found at
[0]).  The model constitution is published as a Microsoft Word document.
In this commit it has been has been converted to markdown.  The text is
unchanged.

[0] https://www.ato.gov.au/businesses-and-organisations/not-for-profit-organisations/statements-and-returns/reporting-requirements-to-self-assess-income-tax-exemption/nfp-self-review-return-question-guide#Prohibitingthedistributionofincomeorasse

[1] https://www.fairtrading.nsw.gov.au/associations-and-co-operatives/associations/starting-an-association/model-constitution