changed prompted by review comments
Changes made:

- Kathy Reid, in PrePur-1: remove superfluous "the".

- Kathy Reid, in AdmCus-1: explicitly allow electronic record keeping.

- Bardi Harborow, in MemReg-5: fixed inconsistency by reverting to the
  original model constitution wording.

- Bardi Harborow, in MetGen-3: Fix trivial inconsistency in _#MetGen-2_.

- Bardi Harborow, after AdmFin-1: Remove the "Note:"

In addition I re-jigged wording on where we must keep our records. The
NSW Model Constitution insisted we keep them "in New South Wales".  This
clashed with us keeping our accounting records in Xero, which is hosted
in AWS in the USA.  I've changed the requirement to "readily available
in Australia".  I hope that works.  I'm not a lawyer.